Create wrapper classes for langchain and api, replace langchain to use streamline and chat model
