+++ main.py
@@ -1,12 +1,11 @@
 import os
-import openai
 import time
 from rich import print
 from dotenv import load_dotenv
-from langchain.llms import OpenAIChat
-from langchain.chains import ConversationChain
-from langchain.chains.conversation.memory import ConversationSummaryBufferMemory
 from tqdm import tqdm
+
+from chat_gpt_wrapper import ChatGpt
+from langchain_wrapper import LangChainGpt
 
 
 def show_waiting_indicator(func):
@@ -23,40 +22,6 @@ def show_waiting_indicator(func):
     return wrapper
 
 
-class ChatGpt:
-    def __init__(self, key, temperature, initial_prompt):
-        openai.api_key = key
-        self.temperature = temperature
-        self.prefix_messages = initial_prompt
-
-    model = "gpt-3.5-turbo"
-
-    def predict(self, prompt):
-        res = openai.ChatCompletion.create(
-            model=self.model,
-            messages=[
-                {"role": "system",
-                 "content": self.prefix_messages},
-                {"role": "user", "content": prompt},
-            ],
-            temperature=self.temperature
-        )
-        return res.choices[0].message.content
-
-
-class LangChainGpt:
-    def __init__(self, initial_content):
-        self.prefix_messages = [{"role": "system",
-                                 "content": initial_content}]
-        self.llm = OpenAIChat(temperature=0, prefix_messages=self.prefix_messages)
-        self.llm_chain = ConversationChain(
-            llm=self.llm,
-            memory=ConversationSummaryBufferMemory(llm=self.llm))
-
-    def predict(self, user_prompt):
-        return self.llm_chain.predict(input=user_prompt)
-
-
 if __name__ == '__main__':
     # Load environment variables
     load_dotenv()
@@ -65,7 +30,7 @@ if __name__ == '__main__':
     initial_prompt = "You are a helpful assistant that is very good at problem solving who thinks step by step."
 
     chat_gpt = ChatGpt(key=token, temperature=1, initial_prompt=initial_prompt)
-    langchain_gpt = LangChainGpt(initial_content=initial_prompt)
+    langchain_gpt = LangChainGpt(system_message=initial_prompt)
 
     command = input("Use langchain? (y/n), q to quit: ")
     while True:
@@ -86,8 +51,9 @@ if __name__ == '__main__':
             continue
 
         if command == "y":
-            answer = langchain_gpt.predict(user_prompt)
-            print(f"\nLangchain: {answer}\n")
+            print(f"\nLangchain:")
+            langchain_gpt.predict(user_prompt)
+            print("\n")
         elif command == "n":
             # print(user_prompt.strip())
             answer = chat_gpt.predict(user_prompt)
